+++ iextrading4j-client/src/main/java/pl/zankowski/iextrading4j/client/rest/request/alternative/CryptoBookRequestBuilder.java
@@ -1,6 +1,6 @@
 package pl.zankowski.iextrading4j.client.rest.request.alternative;
 
-import pl.zankowski.iextrading4j.api.alternative.CryptoBook;
+import pl.zankowski.iextrading4j.api.marketdata.Book;
 import pl.zankowski.iextrading4j.client.rest.manager.RestRequest;
 import pl.zankowski.iextrading4j.client.rest.manager.RestRequestBuilder;
 import pl.zankowski.iextrading4j.client.rest.request.IEXCloudV1RestRequest;
@@ -8,15 +8,15 @@ import pl.zankowski.iextrading4j.client.rest.request.stocks.AbstractStocksReques
 
 import javax.ws.rs.core.GenericType;
 
-public class CryptoBookRequestBuilder extends AbstractStocksRequestBuilder<CryptoBook, CryptoBookRequestBuilder>
-        implements IEXCloudV1RestRequest<CryptoBook> {
+public class CryptoBookRequestBuilder extends AbstractStocksRequestBuilder<Book, CryptoBookRequestBuilder>
+        implements IEXCloudV1RestRequest<Book> {
 
     @Override
-    public RestRequest<CryptoBook> build() {
-        return RestRequestBuilder.<CryptoBook>builder()
+    public RestRequest<Book> build() {
+        return RestRequestBuilder.<Book>builder()
                 .withPath("/crypto/{symbol}/book")
                 .addPathParam(SYMBOL_PARAM_NAME, getSymbol()).get()
-                .withResponse(new GenericType<CryptoBook>() {})
+                .withResponse(new GenericType<Book>() {})
                 .build();
     }