+++ iextrading4j-client/src/test/java/pl/zankowski/iextrading4j/client/rest/request/alternative/CryptoBookRequestBuilderTest.java
@@ -1,8 +1,7 @@
 package pl.zankowski.iextrading4j.client.rest.request.alternative;
 
 import org.junit.Test;
-import pl.zankowski.iextrading4j.api.alternative.CryptoBook;
-import pl.zankowski.iextrading4j.api.alternative.CryptoPrice;
+import pl.zankowski.iextrading4j.api.marketdata.Book;
 import pl.zankowski.iextrading4j.client.rest.manager.MethodType;
 import pl.zankowski.iextrading4j.client.rest.manager.RestRequest;
 
@@ -17,13 +16,13 @@ public class CryptoBookRequestBuilderTest {
     public void shouldSuccessfullyCreateRequest() {
         final String symbol = "BTCUSD";
 
-        final RestRequest<CryptoBook> request = new CryptoBookRequestBuilder()
+        final RestRequest<Book> request = new CryptoBookRequestBuilder()
                 .withSymbol(symbol)
                 .build();
 
         assertThat(request.getMethodType()).isEqualTo(MethodType.GET);
         assertThat(request.getPath()).isEqualTo("/crypto/{symbol}/book");
-        assertThat(request.getResponseType()).isEqualTo(new GenericType<CryptoBook>() {});
+        assertThat(request.getResponseType()).isEqualTo(new GenericType<Book>() {});
         assertThat(request.getPathParams()).containsExactly(entry("symbol", symbol));
         assertThat(request.getQueryParams()).isEmpty();
     }