+++ iextrading4j-test/src/test/java/pl/zankowski/iextrading4j/test/rest/v1/alternative/CryptoBookServiceTest.java
@@ -1,8 +1,8 @@
 package pl.zankowski.iextrading4j.test.rest.v1.alternative;
 
 import org.junit.Test;
-import pl.zankowski.iextrading4j.api.alternative.CryptoBook;
-import pl.zankowski.iextrading4j.api.alternative.CryptoBookEntry;
+import pl.zankowski.iextrading4j.api.marketdata.Book;
+import pl.zankowski.iextrading4j.api.marketdata.BookEntry;
 import pl.zankowski.iextrading4j.client.rest.request.alternative.CryptoBookRequestBuilder;
 import pl.zankowski.iextrading4j.test.rest.v1.BaseIEXCloudV1ServiceTest;
 
@@ -27,20 +27,20 @@ public class CryptoBookServiceTest extends BaseIEXCloudV1ServiceTest {
                         .withHeader("Content-Type", "application/json")
                         .withBodyFile("rest/v1/alternative/CryptoBookResponse.json")));
 
-        final CryptoBook book = cloudClient.executeRequest(new CryptoBookRequestBuilder()
+        final Book book = cloudClient.executeRequest(new CryptoBookRequestBuilder()
                 .withSymbol("BTCUSD")
                 .build());
 
         assertThat(book.getAsks()).hasSize(1);
 
-        final CryptoBookEntry ask = book.getAsks().get(0);
+        final BookEntry ask = book.getAsks().get(0);
         assertThat(ask.getPrice()).isEqualTo(BigDecimal.valueOf(10421));
         assertThat(ask.getSize()).isEqualTo(BigDecimal.valueOf(0.76747926));
         assertThat(ask.getTimestamp()).isEqualTo(1572620748579L);
 
         assertThat(book.getBids()).hasSize(1);
 
-        final CryptoBookEntry bid = book.getBids().get(0);
+        final BookEntry bid = book.getBids().get(0);
         assertThat(bid.getPrice()).isEqualTo(BigDecimal.valueOf(1000));
         assertThat(bid.getSize()).isEqualTo(BigDecimal.valueOf(0.5));
         assertThat(bid.getTimestamp()).isEqualTo(1574688905930L);